Create extension to avoid code duplication
