+++ app/src/main/java/com/infomaniak/drive/data/services/DownloadWorker.kt
@@ -25,22 +25,19 @@ import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import androidx.work.*
-import com.google.gson.reflect.TypeToken
 import com.infomaniak.drive.R
-import com.infomaniak.drive.data.api.ApiRepository
 import com.infomaniak.drive.data.api.ApiRoutes
 import com.infomaniak.drive.data.api.ProgressResponseBody
 import com.infomaniak.drive.data.api.UploadTask
 import com.infomaniak.drive.data.cache.FileController
 import com.infomaniak.drive.data.models.File
 import com.infomaniak.drive.data.models.UserDrive
+import com.infomaniak.drive.extensions.getFileFromRemote
 import com.infomaniak.drive.utils.AccountUtils
 import com.infomaniak.drive.utils.MediaUtils
 import com.infomaniak.drive.utils.MediaUtils.isMedia
 import com.infomaniak.drive.utils.NotificationUtils.downloadProgressNotification
 import com.infomaniak.drive.utils.NotificationUtils.notifyCompat
-import com.infomaniak.lib.core.api.ApiController
-import com.infomaniak.lib.core.models.ApiResponse
 import com.infomaniak.lib.core.networking.HttpClient
 import com.infomaniak.lib.core.networking.HttpUtils
 import com.infomaniak.lib.core.utils.SentryLog
@@ -114,7 +111,12 @@ class DownloadWorker(context: Context, workerParams: WorkerParameters) : Corouti
         if (cacheFile?.exists() == true) cacheFile.delete()
 
 
-        if (file == null || offlineFile == null) getFileFromRemote()
+        if (file == null || offlineFile == null) {
+            getFileFromRemote(fileId, userDrive) { downloadedFile ->
+                file = downloadedFile
+                offlineFile = file?.getOfflineFile(applicationContext, userDrive.driveId)
+            }
+        }
 
         return startOfflineDownload()
     }
@@ -176,26 +178,6 @@ class DownloadWorker(context: Context, workerParams: WorkerParameters) : Corouti
             }
             Result.success()
         } else Result.failure()
-    }
-
-    private fun getFileFromRemote() {
-        val fileDetails = ApiRepository.getFileDetails(File(id = fileId, driveId = userDrive.driveId))
-        val remoteFile = fileDetails.data
-        file = if (fileDetails.isSuccess() && remoteFile != null) {
-            FileController.getRealmInstance(userDrive).use { realm ->
-                FileController.updateExistingFile(newFile = remoteFile, realm = realm)
-            }
-            remoteFile
-        } else {
-            if (fileDetails.error?.exception is ApiController.NetworkException) throw UploadTask.NetworkException()
-
-            val translatedError = fileDetails.translatedError
-            val responseGsonType = object : TypeToken<ApiResponse<File>>() {}.type
-            val translatedErrorText = if (translatedError == 0) "" else applicationContext.getString(translatedError)
-            val responseJson = ApiController.gson.toJson(fileDetails, responseGsonType)
-            throw RemoteFileException("$responseJson $translatedErrorText")
-        }
-        offlineFile = file?.getOfflineFile(applicationContext, userDrive.driveId)
     }
 
     private fun notifyDownloadCancelled() {
